+++ DaresGameBot.Logic/Card.cs
@@ -3,7 +3,7 @@ namespace DaresGameBot.Logic
     public sealed class Card
     {
         public string Description { internal get; set; }
-        public int Players { internal get; set; }
-        public int PartnersToAssign { internal get; set; }
+        public ushort Players { internal get; set; }
+        public ushort PartnersToAssign { internal get; set; }
     }
 }
